services/websocket/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.2
pika==1.3.2
python-dotenv==1.0.0
